Show satisfaction subscores on the land page.
Overall, the graphs now show that at first population rises, but then quality of life starts to drop as too many people become difficult to accommodate in a fully common participatory ritual. Eventually this causes people to move out and population crashes.

This is a simplification, as interaction stress and cultural preference for smaller settlement sizes would be major factors as well, but for now it implements the idea that disagreements become a problem in large, unstructured societies. The next step, therefore, is to implement ritual structure changes and other ways for clans to deal with ritual coordination problems.