Add ritual structure option to exclude clans that are messing up the rituals.
It was noticed that sometimes ritual quality was low at the start because of one or two clans with much lower effectiveness than others.

The effects will probably need some modification, as whether or not the clans want to be cut out would have a lot of bearing on what happens.