Fix problems with clans not migrating as expected.
The immediate problem was that a limit intending to prevent more than a fraction of each settlement from moving in a turn (so everyone doesn't move away at once due to crowding) was being applied region-wide instead.

But also, the code to move based on local goods availability wasn't working very well, in part because it's hard to get an accurate estimate of goods at a new settlement. For now, we use land availability instead, which is stabler.