Add a labor planning report to clan details.
This finally makes it easy to track turn to turn what a given clan is doing for labor allocations. This showed that as populations grow and fishing returns decline, clans initially may shift to more farming, but then they shift back more to fishing because contentment is still low, but food quality is an issue. Conclusion is that we need to treat fish as more of a superior good: in other words, hungry clans will take what they can get.